# access logs: perf: cache the formatted date/time