--- a/test/common/common/utility_speed_test.cc
+++ b/test/common/common/utility_speed_test.cc
@@ -31,6 +31,11 @@ static void BM_AccessLogDateTimeFormatter(benchmark::State& state) {
   static std::mt19937 prng(1); // PRNG with a fixed seed, for repeatability
   static std::uniform_int_distribution<long> distribution(-10, 20);
   for (auto _ : state) {
+    // TODO(brian-pane): The next line, which computes the next input timestamp,
+    // currently accounts for ~30% of the CPU time of this benchmark test. If
+    // the AccessLogDateTimeFormatter implementation is optimized further, we
+    // should precompute a sequence of input timestamps so the benchmark's own
+    // overhead won't obscure changes in the speed of the code being benchmarked.
     time += std::chrono::milliseconds(static_cast<int>(distribution(prng)));
     outputBytes += Envoy::AccessLogDateTimeFormatter::fromTime(time).length();
   }