*Description*:
Add thread-local caching of the formatted date/time for the access logger.

In local testing with production traffic, this reduces total Envoy CPU usage by ~1%.

*Risk Level*: Low

*Testing*:
New unit tests included

*Docs Changes*:
N/A

*Release Notes*:
N/A

Addresses issue #2879

Signed-off-by: Brian Pane <bpane@pinterest.com>