@mattklein123 2018-04-04
> Thanks. Can you quantify the performance benefits in the commit message? I think where we're adding complexity for performance, we should generally include some idea of what we're buying.

Optimally, via a benchmark with the BM stuff ...