@brian-pane 2018-04-04
I'll add a benchmark test.