@brian-pane 2018-04-04
Here's the result of the benchmark test I just added, for before and after applying the optimization from this PR:
```
-------------------------------------------------------------------------------------
Benchmark                                              Time           CPU Iterations
-------------------------------------------------------------------------------------
Before:
BM_AccessLogDateTimeFormatter                    1190942 ns    1190847 ns        588
After:
BM_AccessLogDateTimeFormatter                     703801 ns     703793 ns        995
```
Each iteration in that benchmark makes 1,000 calls  to `AccessLogDateTimeFormatter::fromTime`, with varying inputs to exercise the cache-hit and cache-miss cases.