@htuch 2018-04-05
I think we should be able to make this much faster if we're engaging in this micro-optimization. If we're just rewriting the last 3 digits of an existing string, it should be possible to do this in a few hundred cycles tops. Do you think you could do something like this?

It's kind of scary doing this low-level optimization without having a macro-level objective as well. Based on your earlier perf result in the linked issue this is only 1% of overall time. Are there situations where this is more acute?