@brian-pane 2018-04-05
My macro-level objective is to fix the top inefficiencies that collectively account for 10+% of Envoy’s CPU usage. That list includes the ioctl before reads, the http_parser, the access logger in general, and this time formatter. Most of those are 1-2% wins individually, which is consistent with other proxies I’ve profiled at a similar point in their development.

Anyway, I’ll do some more optimization on this diff to further increase the speedup.