@mattklein123 2018-04-05
> My macro-level objective is to fix the top inefficiencies that collectively account for 10+% of Envoy’s CPU usage. That list includes the ioctl before reads, the http_parser, the access logger in general, and this time formatter.

This is probably obvious, but I would also clarify that this 10% is very specific to your deployment. It's fine to do perf optimizations, but we should definitely also quantify the deployment types in which they matter. (In your case HTTP/1.1 with heavy text log usage I guess).