@brian-pane 2018-04-09
Latest benchmark results:
```
-------------------------------------------------------------------------------------
Benchmark                                              Time           CPU Iterations
-------------------------------------------------------------------------------------
Before:
BM_AccessLogDateTimeFormatter                       1309 ns       1309 ns     533858
After:
BM_AccessLogDateTimeFormatter                        277 ns        277 ns    2544025
```
Notes:
* With the latest updates to this PR, each iteration contains 1 call to `AccessLogDateTimeFormatter::fromTime`, instead of the 1000 calls it was doing previously.
* The benchmark understates the speedup, because ~80 ns of both the "before" and the "after" measurements is spent computing the next input before calling the function being benchmarked. I.e., the overhead of the test itself is a big fraction of the "after" value. I don't think that's worth fixing right now, but I added a TODO comment for it.