@brian-pane 2018-04-10
With these latest changes, the speed has improved a bit more:
```
-------------------------------------------------------------------------------------
Benchmark                                              Time           CPU Iterations
-------------------------------------------------------------------------------------
Original:
BM_AccessLogDateTimeFormatter                       1309 ns       1309 ns     533858
With cached string:
BM_AccessLogDateTimeFormatter                        277 ns        277 ns    2544025
With cached character array:
BM_AccessLogDateTimeFormatter                        165 ns        165 ns    4177561
```